chore: rename leaderboardDef folder to leaderboard
